+++ src/observable/observable_err.rs
@@ -72,7 +72,7 @@ where
       is_stopped: false,
       marker: TypeHint::new(),
     }));
-    unsub.request(u128::MAX);
+    unsub.request(usize::MAX);
     SubscriptionWrapper(unsub)
   }
 }
