+++ src/subject.rs
@@ -13,11 +13,11 @@ use std::sync::{Arc, Mutex};
 pub struct Subject<V, S> {
   pub(crate) observers: SubjectObserver<V>,
   pub(crate) subscription: S,
-  pub(crate) last_requested: u128,
+  pub(crate) last_requested: usize,
 }
 
 impl<O, U: SubscriptionLike> SubscriptionLike for Subject<O, U> {
-  fn request(&mut self, requested: u128) {
+  fn request(&mut self, requested: usize) {
     self.last_requested += requested;
   }
